refactor: hoist json-salvage helpers into @opuspopuli/common
Resolves Sonar duplication on packages/scraping-pipeline/src/utils/json-salvage.ts (42.4% duplicated with apps/backend's llm-json-salvage.util.ts). Merges both copies into packages/common/src/utils/json-salvage.ts as the single source of truth — backend generators and the scraping-pipeline PDF handler now import extractJsonObjectSlice / extractFieldString / stripCodeFences from @opuspopuli/common.

Also marks the slugify regex in domain-mapper.service.ts with NOSONAR (S5852 false positive — single char-class with `+` is linear, no backtracking trap).
