+++ apps/backend/src/apps/region/src/domains/committee-summary-generator.service.ts
@@ -1,12 +1,13 @@
 import { Inject, Injectable, Logger, Optional } from '@nestjs/common';
 import { ConfigService } from '@nestjs/config';
 import { PromptClientService } from '@opuspopuli/prompt-client';
-import type { CommitteeAssignment, ILLMProvider } from '@opuspopuli/common';
-import { DbService } from '@opuspopuli/relationaldb-provider';
 import {
   extractFieldString,
   extractJsonObjectSlice,
-} from './llm-json-salvage.util';
+  type CommitteeAssignment,
+  type ILLMProvider,
+} from '@opuspopuli/common';
+import { DbService } from '@opuspopuli/relationaldb-provider';
 
 /** Minimal shape of a representative needed to render a committee summary. */
 interface RepForSummary {
