+++ packages/common/src/utils/index.ts
@@ -1,1 +1,6 @@
 export { batchTransaction } from "./batch-transaction.js";
+export {
+  extractFieldString,
+  extractJsonObjectSlice,
+  stripCodeFences,
+} from "./json-salvage.js";
